+++ R/datatableInternal.R
@@ -16,15 +16,15 @@
 
 datatableInternal <- function(x,
                               delim = "\n",
-                              style = "default",
+                              style = styleDT(),
                               caption = NULL,
-                              groupColumn = NULL) {
+                              groupColumn = NULL,
+                              groupOrder = NULL) {
 
   # Package checks
   rlang::check_installed("DT")
   rlang::check_installed("htmltools")
 
-  style <- validateStyle(style, "datatable")
   options <- style[c(
     "scrollX", "scrollY", "scrollCollapse", "pageLength", "lengthMenu",
     "searchHighlight", "scroller", "deferRender", "fixedColumns",
@@ -37,13 +37,27 @@ datatableInternal <- function(x,
   colnames(x) <- gsub("\\[header\\]|\\[header_level\\]|\\[header_name\\]|\\[column_name\\]", "", colnames(x))
 
   # groupColumn
-  if (length(groupColumn) > 0) {
+  if (length(groupColumn[[1]]) != 0) {
     nameGroup <- names(groupColumn)
     x <- x |>
       tidyr::unite(
         !!nameGroup, groupColumn[[1]], sep = "; ", remove = TRUE, na.rm = TRUE
-      ) |>
-      dplyr::relocate(!!nameGroup)
+      )
+    groupLevel <- unique(x[[nameGroup]])
+    if (!is.null(groupOrder)) {
+      if (any(!groupLevel %in% groupOrder)) {
+        cli::cli_abort(c(
+          "x" = "`groupOrder` supplied does not macth the group variable created based on `groupName`.",
+          "i" = "Group variables to use in `groupOrder` are the following: {groupLevel}"
+        ))
+      } else {
+        groupLevel <- groupOrder
+      }
+    }
+    x <- x |>
+      dplyr::mutate(!!nameGroup := factor(.data[[nameGroup]], levels = groupLevel)) |>
+      dplyr::arrange_at(nameGroup) |>
+      dplyr::relocate(dplyr::all_of(nameGroup))
     options$rowGroup = list(dataSrc = 0)
   }